+++ lib/allthenews_app.dart
@@ -1,5 +1,5 @@
 import 'package:allthenews/ui/common/theme/theme.dart';
-import 'package:allthenews/ui/pages/news_list/news_list_page.dart';
+import 'package:allthenews/ui/pages/news_list/home_page.dart';
 import 'package:flutter/material.dart';
 
 class AllTheNewsApp extends StatelessWidget {
@@ -7,7 +7,7 @@ class AllTheNewsApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       theme: newsTheme,
-      home: NewsListPage(),
+      home: HomePage(),
     );
   }
 }
