+++ lib/ui/pages/news_list/primary_news/primary_news_list_entity.dart
@@ -1,81 +1,103 @@
 class PrimaryNewsListEntity {
   final String imageUrl;
   final String title;
+  final String date;
+  final String time;
   final String authorName;
   final String authorImageUrl;
-  final String url;
+  final String articleUrl;
 
   PrimaryNewsListEntity({
     this.imageUrl,
     this.title,
+    this.date,
+    this.time,
     this.authorName,
     this.authorImageUrl,
-    this.url,
+    this.articleUrl,
   });
 }
 
 final primaryNewsListEntities = <PrimaryNewsListEntity>[
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
-    title: "Something wrong is happening with the newest flutter build",
+    title: "Something wrong is happening with ",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
     title: "New laptops seem to be stuck to the board",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
     title: "Huawei Loses Millions Of Users As Serious",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
     title: "Samsung Surprise As World’s First Smartphone ",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
     title: "It’s ‘Comet SWAN’ Week. ",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
     title: "Summer Travel To Europe",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
     title: "‘White Lines’ Review: Netflix’s Murder Mystery In Ibiza",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
     title: "What’s Going On With Comet SWAN? ",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   PrimaryNewsListEntity(
     imageUrl: 'https://i.picsum.photos/id/9/800/800.jpg',
     title: "Hackers Claim To Have Trump’s Dirty Laundry",
+    date: 'Sept. 26, 2019',
+    time: '1:41 PM',
     authorName: "Marek Aureliusz",
     authorImageUrl: 'https://www.ludoviccareme.com/files/image_211_image_fr.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
 ];
