+++ lib/ui/pages/news_list/primary_news/primary_news_list_item.dart
@@ -52,7 +52,7 @@ class PrimaryNewsListItem extends StatelessWidget {
                     onTap: () => Navigator.push(
                       context,
                       MaterialPageRoute(
-                        builder: (context) => WebViewPage(url: news.url),
+                        builder: (context) => WebViewPage(url: news.articleUrl),
                       ),
                     ),
                   ),
