+++ lib/ui/pages/news_list/secondary_news/secondary_news_list_entity.dart
@@ -2,14 +2,14 @@ class SecondaryNewsListEntity {
   final String title;
   final String date;
   final String time;
-  final String url;
+  final String articleUrl;
   final String imageUrl;
 
   SecondaryNewsListEntity({
     this.title,
     this.date,
     this.time,
-    this.url,
+    this.articleUrl,
     this.imageUrl,
   });
 }
@@ -20,62 +20,62 @@ final secondaryNewsListEntities = <SecondaryNewsListEntity>[
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   SecondaryNewsListEntity(
     title: 'A faint hope emerges as people return to public places.',
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   SecondaryNewsListEntity(
     title: 'Britain’s new guidelines leave more questions than answers.',
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   SecondaryNewsListEntity(
     title: 'Consumers are not optimistic about home prices or keeping their jobs.',
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   SecondaryNewsListEntity(
     title: 'Natural gas exports slow as the pandemic reduces global demand.',
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   SecondaryNewsListEntity(
     title: 'The airline business is terrible. It will probably get worse.',
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   SecondaryNewsListEntity(
     title: 'N.Y.C. will likely stay shut down until June, the mayor said.',
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   SecondaryNewsListEntity(
     title: 'In New York City, 38 children have become ill from a virus-linked syndrome.',
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
   SecondaryNewsListEntity(
     title: 'In New Jersey, 72 people have died at a home for veterans.',
     date: 'Sept. 26, 2019',
     time: '1:41 PM',
     imageUrl: 'https://i.picsum.photos/id/9/250/250.jpg',
-    url: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
+    articleUrl: 'https://www.nytimes.com/2020/05/14/obituaries/jonathan-adewumi-dies-coronavirus.html',
   ),
 ];
