+++ lib/ui/pages/news_list/secondary_news/secondary_news_list_item.dart
@@ -29,7 +29,7 @@ class SecondaryNewsListItem extends StatelessWidget {
         onTap: () => Navigator.push(
           context,
           MaterialPageRoute(
-            builder: (context) => WebViewPage(url: news.url),
+            builder: (context) => WebViewPage(url: news.articleUrl),
           ),
         ),
         child: Padding(
